refactor: fix loader in reverse swap
